Abstract auth0Client into a generic AuthClient implementation
