+++ quadratic-client/src/app/web-workers/pythonWebWorker/pythonWebWorker.ts
@@ -1,6 +1,6 @@
 import { events } from '@/app/events/events';
 import { LanguageState } from '@/app/web-workers/languageTypes';
-import { authClient } from '@/auth';
+import { authClient } from '@/auth/auth';
 import mixpanel from 'mixpanel-browser';
 import { quadraticCore } from '../quadraticCore/quadraticCore';
 import { ClientPythonMessage, PythonClientGetJwt, PythonClientMessage } from './pythonClientMessages';
